feat: make command as clazz
