+++ register.js
@@ -1,1 +1,2 @@
-require('./dist/register');
+const Register = require('./dist/register').default;
+new Register().init();
