+++ tsconfig.json
@@ -3,7 +3,7 @@
     "strict": true,
     "target": "es2017",
     "module": "commonjs",
-    "moduleResolution": "node",
+    "moduleResolution": "nodenext",
     "outDir": "dist",
     "noImplicitAny": false,
     "allowSyntheticDefaultImports": true,
